ai-chat-python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.24.4
pandas==1.5.3
pandasai==3.0.0b14
pandasai-litellm==0.0.1
python-dotenv==1.0.1
supabase==1.2.0
matplotlib==3.5.1
gunicorn==21.2.0
pyyaml==6.0.1
beautifulsoup4==4.12.2 
supafunc==0.2.2
polars==0.19.19
seaborn==0.12.2
duckdb==0.9.2 